use shared parse UA logic
